refact: unify every parts
