+++ Augmentation.py
@@ -4,6 +4,7 @@ import os
 
 from colorama import Fore, Style
 from PIL import Image, ImageEnhance, ImageFilter
+from utils import is_jpg
 
 g_suffixes = {
     "_blurred",
@@ -13,19 +14,6 @@ g_suffixes = {
     "_rotated",
     "_projected"
 }
-
-
-def is_jpg(file_path: str) -> bool:
-    """
-    Check if the given file is a JPEG image (.jpg or .jpeg), case-insensitive.
-
-    :param file_path: Path to the file to check.
-    :return: True if the file exists
-        and ends with .jpg or .jpeg, False otherwise.
-    """
-    return (os.path.isfile(file_path)
-            and file_path.lower().endswith(
-                ('.jpg', '.jpeg')))
 
 
 def zoom(img, zoom_factor):
